+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ### Added
 
+- Added placeholder text for Units without Exercises
+
 ### Changed
 
 ### Fixed
 
 ### Removed
-